+++ website_multi_company_portal/__manifest__.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 {
-    "name": """Real Multi Website (eCommerce extension)""",
-    "summary": """Multi Website support in eCommerce""",
-    "category": "eCommerce",
+    "name": """Real Multi Website (portal extension)""",
+    "summary": """Multi Website support in Portal""",
+    "category": "Portal",
     "live_test_url": "http://apps.it-projects.info/shop/product/website-multi-company?version=10.0",
     "images": [],
-    "version": "1.0.1",
+    "version": "1.0.0",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",
@@ -17,11 +17,10 @@
 
     "depends": [
         "website_multi_company",
-        "website_sale",
+        "website_portal",
     ],
     "external_dependencies": {"python": [], "bin": []},
     "data": [
-        "views/product_public_category_views.xml",
     ],
     "qweb": [
     ],