+++ website_multi_company/doc/changelog.rst
@@ -1,3 +1,19 @@
+`1.2.3`
+-------
+
+- **FIX:** fix multi-theme for themes that have theme-like dependencies (problem with views that have position="replace")
+
+`1.2.2`
+-------
+
+- **FIX:** fix multi-theme for themes that have theme-like dependencies
+
+`1.2.0`
+-------
+
+- **FIX:** allow install several themes at once
+- **ADD:** support of custom themes
+
 `1.1.0`
 -------